+++ lib/airbrake/rails.rb
@@ -79,6 +79,7 @@ module Airbrake
         require 'airbrake/rails/curb' if defined?(Curl) && defined?(Curl::CURB_VERSION)
         require 'airbrake/rails/http' if defined?(HTTP) && defined?(HTTP::Client)
         require 'airbrake/rails/http_client' if defined?(HTTPClient)
+        require 'airbrake/rails/typhoeus' if defined?(Typhoeus)
 
         if defined?(Excon)
           require 'airbrake/rails/excon_subscriber'