+++ spec/apps/rails/dummy_app.rb
@@ -42,6 +42,7 @@ class DummyApp < Rails::Application
     get '/breakdown_excon' => 'dummy#breakdown_excon'
     get '/breakdown_http_rb' => 'dummy#breakdown_http_rb'
     get '/breakdown_http_client' => 'dummy#breakdown_http_client'
+    get '/breakdown_typhoeus' => 'dummy#breakdown_typhoeus'
     get '/notify_airbrake_helper' => 'dummy#notify_airbrake_helper'
     get '/notify_airbrake_sync_helper' => 'dummy#notify_airbrake_sync_helper'
     get '/active_record_after_commit' => 'dummy#active_record_after_commit'
@@ -121,7 +122,8 @@ class DummyController < ActionController::Base
       'dummy/breakdown_curl_http_multi.html.erb' => 'breakdown_curl_http_multi',
       'dummy/breakdown_excon.erb' => 'breakdown_excon',
       'dummy/breakdown_http_rb.erb' => 'breakdown_http_rb',
-      'dummy/breakdown_http_client.erb' => 'breakdown_http_client'
+      'dummy/breakdown_http_client.erb' => 'breakdown_http_client',
+      'dummy/breakdown_typhoeus.erb' => 'breakdown_typhoeus'
     )
   ]
 
@@ -174,6 +176,11 @@ class DummyController < ActionController::Base
   def breakdown_http_client
     HTTPClient.new.get('http://example.com')
     render 'dummy/breakdown_http_client.html.erb'
+  end
+
+  def breakdown_typhoeus
+    Typhoeus.get('example.com')
+    render 'dummy/breakdown_typhoeus.html.erb'
   end
 
   def notify_airbrake_helper